+++ packages/rtk-query-codegen/package.json
@@ -2,17 +2,22 @@
   "name": "@layer5/rtk-query-codegen",
   "version": "0.0.1",
   "description": "Codegen to create rtk-query api from openapi schema",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "bin": "./app.js",
+  "license": "ISC",
   "author": "",
+  "main": "./dist/index.js",
+  "bin": "./dist/index.js",
+  "scripts": {
+    "api-generate": " npx @rtk-query/codegen-openapi openapi-config.ts",
+    "build": "tsc -p ."
+  },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.6",
     "@rtk-query/codegen-openapi": "^1.0.0",
-    "commander": "^11.0.0",
-    "prettier-plugin-organize-imports": "^3.2.3",
-    "ts-node": "^10.9.1"
+    "commander": "^11.0.0"
+  },
+  "devDependencies": {
+    "@types/node": "^20.7.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
 }
